requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
psycopg2-binary==2.9.1
Pillow==8.3.1
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.0.1
email-validator==1.1.3
Flask-Bcrypt==0.7.1